+++ sentry/src/main/java/io/sentry/protocol/Contexts.java
@@ -72,7 +72,7 @@ public class Contexts implements JsonSerializable {
     return toContextType(SpanContext.TYPE, SpanContext.class);
   }
 
-  public void setTrace(final @Nullable SpanContext traceContext) {
+  public void setTrace(final @NotNull SpanContext traceContext) {
     Objects.requireNonNull(traceContext, "traceContext is required");
     this.put(SpanContext.TYPE, traceContext);
   }