+++ sentry/src/test/java/io/sentry/CombinedContextsViewTest.kt
@@ -37,7 +37,7 @@ class CombinedContextsViewTest {
     fun `uses default context CURRENT`() {
         fixture.getSut()
         val combined = CombinedContextsView(fixture.global, fixture.isolation, fixture.current, ScopeType.CURRENT)
-        combined.trace = SpanContext("some")
+        combined.setTrace(SpanContext("some"))
         assertEquals("some", fixture.current.trace?.op)
     }
 
@@ -45,7 +45,7 @@ class CombinedContextsViewTest {
     fun `uses default context ISOLATION`() {
         fixture.getSut()
         val combined = CombinedContextsView(fixture.global, fixture.isolation, fixture.current, ScopeType.ISOLATION)
-        combined.trace = SpanContext("some")
+        combined.setTrace(SpanContext("some"))
         assertEquals("some", fixture.isolation.trace?.op)
     }
 
@@ -53,16 +53,16 @@ class CombinedContextsViewTest {
     fun `uses default context GLOBAL`() {
         fixture.getSut()
         val combined = CombinedContextsView(fixture.global, fixture.isolation, fixture.current, ScopeType.GLOBAL)
-        combined.trace = SpanContext("some")
+        combined.setTrace(SpanContext("some"))
         assertEquals("some", fixture.global.trace?.op)
     }
 
     @Test
     fun `prefers trace from current context`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
-        fixture.isolation.trace = SpanContext("isolation")
-        fixture.global.trace = SpanContext("global")
+        fixture.current.setTrace(SpanContext("current"))
+        fixture.isolation.setTrace(SpanContext("isolation"))
+        fixture.global.setTrace(SpanContext("global"))
 
         assertEquals("current", combined.trace?.op)
     }
@@ -70,8 +70,8 @@ class CombinedContextsViewTest {
     @Test
     fun `uses isolation trace if current context does not have it`() {
         val combined = fixture.getSut()
-        fixture.isolation.trace = SpanContext("isolation")
-        fixture.global.trace = SpanContext("global")
+        fixture.isolation.setTrace(SpanContext("isolation"))
+        fixture.global.setTrace(SpanContext("global"))
 
         assertEquals("isolation", combined.trace?.op)
     }
@@ -79,7 +79,7 @@ class CombinedContextsViewTest {
     @Test
     fun `uses global trace if current and isolation context do not have it`() {
         val combined = fixture.getSut()
-        fixture.global.trace = SpanContext("global")
+        fixture.global.setTrace(SpanContext("global"))
 
         assertEquals("global", combined.trace?.op)
     }
@@ -87,7 +87,7 @@ class CombinedContextsViewTest {
     @Test
     fun `sets trace on default context`() {
         val combined = fixture.getSut()
-        combined.trace = SpanContext("some")
+        combined.setTrace(SpanContext("some"))
 
         assertNull(fixture.current.trace)
         assertEquals("some", fixture.isolation.trace?.op)
@@ -414,7 +414,7 @@ class CombinedContextsViewTest {
     @Test
     fun `size combines contexts`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
+        fixture.current.setTrace(SpanContext("current"))
         fixture.isolation.setApp(App().also { it.appName = "isolation" })
         fixture.global.setGpu(Gpu().also { it.name = "global" })
 
@@ -424,9 +424,9 @@ class CombinedContextsViewTest {
     @Test
     fun `size considers overrides`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
-        fixture.isolation.trace = SpanContext("isolation")
-        fixture.global.trace = SpanContext("global")
+        fixture.current.setTrace(SpanContext("current"))
+        fixture.isolation.setTrace(SpanContext("isolation"))
+        fixture.global.setTrace(SpanContext("global"))
 
         assertEquals(1, combined.size)
     }
@@ -440,7 +440,7 @@ class CombinedContextsViewTest {
     @Test
     fun `isNotEmpty if current has value`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
+        fixture.current.setTrace(SpanContext("current"))
 
         assertFalse(combined.isEmpty)
     }
@@ -470,28 +470,28 @@ class CombinedContextsViewTest {
     @Test
     fun `containsKey current`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
+        fixture.current.setTrace(SpanContext("current"))
         assertTrue(combined.containsKey("trace"))
     }
 
     @Test
     fun `containsKey isolation`() {
         val combined = fixture.getSut()
-        fixture.isolation.trace = SpanContext("isolation")
+        fixture.isolation.setTrace(SpanContext("isolation"))
         assertTrue(combined.containsKey("trace"))
     }
 
     @Test
     fun `containsKey global`() {
         val combined = fixture.getSut()
-        fixture.global.trace = SpanContext("global")
+        fixture.global.setTrace(SpanContext("global"))
         assertTrue(combined.containsKey("trace"))
     }
 
     @Test
     fun `keys combines contexts`() {
         val combined = fixture.getSut()
-        fixture.current.trace = SpanContext("current")
+        fixture.current.setTrace(SpanContext("current"))
         fixture.isolation.setApp(App().also { it.appName = "isolation" })
         fixture.global.setGpu(Gpu().also { it.name = "global" })
 
@@ -502,7 +502,7 @@ class CombinedContextsViewTest {
     fun `entrySet combines contexts`() {
         val combined = fixture.getSut()
         val trace = SpanContext("current")
-        fixture.current.trace = trace
+        fixture.current.setTrace(trace)
         val app = App().also { it.appName = "isolation" }
         fixture.isolation.setApp(app)
         val gpu = Gpu().also { it.name = "global" }